data: withReverseList_template
